pyproject.toml
[tool.poetry]
name = "oc_chess_club"
version = "0.1.0"
description = ""
authors = ["PabloLec <pablo.lecolinet@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
tinydb = "^4.4.0"
typer = "^0.3.2"
colorama = "^0.4.4"
shellingham = "^1.4.0"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
mkdocs-autorefs = "^0.1.1"
mkdocs-gen-files = "^0.3.1"
mkdocs-material = "^7.1.1"
mkdocstrings = "^0.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
